net/vmxnet3: reallocate shared memzone on re-config
When adding a DPDK port to a bridge using ovs-vswitchd with DPDK,
the vmxnet3 device fails to activate due to mismatched magic number.
Doing this will incur the following operations: start the port,
stop the port, reconfigure and re-start the port.  The reconfig
could request different number of tx/rx queues but the driver
bypasses allocating the queuedesc memzone of proper size (which
depends on the queue size) if there is an existing one.  This
results in a memzone with wrong size and potential invalid memory
access. To fix this, this change will free the memzone if found
and reserve a new one.

Signed-off-by: Yong Wang <yongwang@vmware.com>
Reviewed-by: Guolin Yang <gyang@vmware.com>
Reviewed-by: Daniele Di Proietto <ddiproietto@vmware.com>
Tested-by: Daniele Di Proietto <ddiproietto@vmware.com>