testpmd: fix fdir command on MAC and tunnel modes
The flow_director_filter commands has a pf|vf option for most modes
except for MAC-VLAN and tunnel. On Intel NIC's these modes are not
supported under virtualized environments.
But the application was checking that this field was parsed for these cases,
even though this token is not registered with the cmdline parser.

This patch skips checking of this field for the commands that don't
accept it.

Signed-off-by: Frederico Cadete <Frederico.Cadete-ext@oneaccess-net.com>