add mtu set in virtio
This functionality is required mostly in the cloud infrastructure.
For example, if we use gre or vxlan network between compute and
controller, then we should not use 1500 mtu in the guest as with
encapsulation the sixe of the packet will be more and will get
dropped in the infrastructure. So, in that case we should honor
the mtu size sent by the dhcp server and configure the same on
the virtual interfaces in the guest. This will also keep a
consistent mtu through out the infrastructure.

souvikdey33 (1):
  Signed-off-by: Souvik Dey <sodey@sonusnet.com>

 drivers/net/virtio/virtio_ethdev.c | 12 ++++++++++++
 1 file changed, 12 insertions(+)

--
From 2e4b391fe90ba5e617611e341a7d260dd3dd9144 Mon Sep 17 00:00:00 2001
From: souvikdey33 <sodey@sonusnet.com>
Date: Fri, 26 Aug 2016 20:46:21 -0400
Subject: [PATCH v1] Signed-off-by: Souvik Dey <sodey@sonusnet.com>

Fixes: 1fb8e8896ca8 ("Signed-off-by: Souvik Dey <sodey@sonusnet.com>")

Virtio interfaces should also support setting of mtu, as in case of cloud
it is expected to have the consistent mtu across the infrastructure that
the dhcp server sends and not hardcoded to 1500(default).