net/ixgbe/base: support X550em_a SGMII FC autoneg
Create FC (flow control) setup and enable FC function with code to
configure the PHY registers to advertize FC settings according to the
user requested settings and to negotiate the same with the link partner
and arrive at a negotiated setting (RX and TX enabled, TX only enabled,
FC disabled and RX enabled (which also enables TX) ). Remove the part of
code that performs FC auto-negotiation (AN) in ixgbe_setup_m88 and put it
in a new function to maintain consistency with other FC AN implementations
so far.  Add function pointers to the setup-fc and enable-fc functions so
that they get called during init to perform FC AN.

Signed-off-by: Xiao Wang <xiao.w.wang@intel.com>