net/ixgbe/base: hold semaphore for shadow RAM access
The semaphore is not being held for complete shadow RAM accesses
which could result in corruption. Refactor the code so that it is
possible to hold the semaphore around ixgbe_host_interface_command
by introducing an unlocked form. This patch also eliminates the
function ixgbe_read_ee_hostif_data_X550 in favor of the function
ixgbe_read_ee_hostif_X550 and it now gets both semaphore bits
at once instead of nesting them. The new arrangement is able to
get both the management interface and the EEPROM semaphores at the
same time instead of separately.

Signed-off-by: Xiao Wang <xiao.w.wang@intel.com>