examples/l3fwd-power: fix inappropiate Rx queue count method
Since commit b4f3c136a179 ("net/ixgbe: support checksum flags in SSE vector Rx"),
when HW IP Checksum is enabled, the RX vector function is available to use.
In case of L3fwd-power, since it is enabled, the RX vector function
was not used, but after this commit, it is. This has affected the way
to calculate how full an RX queue is, using rte_eth_rx_descriptor done
function, making the frequency to be at maximum, even when the frequency
is very low.

This commit reverts the way to know how full a queue is to the
previous way, using rte_eth_rx_queue_count(), making it work for both
vector and scalar RX functions.

Fixes: b451aa39db31 ("examples/l3fwd-power: use DD bit rather than RX queue count")

Signed-off-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>