eventdev: header and implementation
This commit adds the eventdev API header, and the library infrastructure
necessary to work with an eventdev instance.  The header is mostly the same
as discussed on the mailing list, in particular see the libeventdev RFC
thread[1].  A second thread[2] gathered input from the community, which
resulted in various modifications to the header.

Since then, some modifications have been made that were discovered during
the implementation of the software implementation of an eventdev. These
changes are present in this header as suggestions, and feedback on these
changes is welcomed. Some of the main items are:
   - A statistics API, similar to rte_ethdev_stats_get()
   - A dump API, similar to rte_ring_dump()
   - Return value of rte_eventdev_enqueue() changed for dequeue consistency
   - rte_event has implementation/operation field as discussed[3]
   - Add enqueue types using operation[4]

This commit does not change any existing DPDK code - only adds the eventdev
header and infrastructure.

Signed-off-by: Gage Eads <gage.eads@intel.com>
Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>

[1] http://dpdk.org/ml/archives/dev/2016-August/045181.html
[2] http://dpdk.org/ml/archives/dev/2016-October/048196.html
[3] http://dpdk.org/ml/archives/dev/2016-October/049459.html
[4] http://dpdk.org/ml/archives/dev/2016-October/049641.html