event/sw: software eventdev implementation
This commit adds a software implementation of the eventdev API. The
implementation here is intended to enable the community to use the eventdev
API, and test if the API serves the purpose that it is designed to. It
should be noted this is an RFC implementation, and hence there should be no
performance expectations. Note that the code added here is based on a
prototype implementation, and hence some cleanup is expected to be
necessary.

The main components of the implementation is three files:
  - sw_evdev.c              Creation, configuration, etc
  - sw_evdev_worker.c       Worker cores' enqueue (etc) functions
  - sw_evdev_scheduler.c    Core pkt scheduling implementation

This commit only adds the implementation, no existing DPDK files are modified.

Signed-off-by: Gage Eads <gage.eads@intel.com>
Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>