lpm: rte_lpm_iterate() - iterate through the routes
In practice, there's a need to iterate through the entries
of a rte_lpm, apart from the usual insert/delete/lookup
operations.  This is useful for debugging and perhaps for
things like removing all entries referencing a certain nexthop.

This patch implements this through rte_lpm_iterate(), which
uses a cursor (or iterator) to keep track of the current
position.  Client code doesn't need to be aware of rte_lpm
implementation details.

Change-Id: I28ea3d7d92f318988444553ee2bb30b709bcb3b6
Signed-off-by: Jorgen Grahn <jorgen.grahn@hiq.se>
Signed-off-by: alloc <alloc.young@gmail.com>