net/mlx5: fix link status is always inconsist
Query the link status can end up in an inconsist state where the port is
down but it is reporting speed. For that another query is scheduled for a
later time.
A race condition is possible between the scheduled call and other link
status interrupt handlers. When the scheduled query by-pass those handlers,
the link status will be stuck in an in-consist state.
This patch addresses the race condition by not blocking link status
queries in case delayed query is on the flight.

Fixes: 198a3c339a8f ("mlx5: handle link status interrupts")
CC: stable@dpdk.org

Signed-off-by: Shahaf Shuler <shahafs@mellanox.com>