ring: allow macros to work with any type of object
Modify the enqueue and dequeue macros to support copying any type of
object by passing in the exact object type. We no longer need a
placeholder element array in the ring structure, since the macros just
take the address of the end of the structure, so remove it, leaving
the rte_ring structure as a ring header only.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>