doc: use corelist instead of coremask
The coremask option in DPDK is difficult to use and we should be
promoting the use of the corelist (-l) option. The patch
adjusts the docs to use -l EAL option instead of the -c option.

The patch only changes the docs and not the code as the -c option
will continue to exist unless it is removed in the future. The -c
option should be kept to maintain backward compatibility.

v3 - squash all of the changes into a single commit.
v2 - reset changes to taskset back to using -c.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>