rte, eal: Rename pci_update_device and make it public
The reaon is that sometimes we only like to rebound the
kernel driver or VFIO or UIO or other drivers for
this device after rte_eal_detach function. Function
rte_eal_pci_probe_one not only updates the device but
also probes the rte_eal_driver for this device, it is
not flexible.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>