test: move tests to separate folder
Tests are part of app folder and compiled with library every time.
Moving tests into a "test" folder which won't be compiled by default.
To compile tests, need to give explicit "make test" command.

"make test" was previously used to run tests, which renamed to
"make test_run" with this patch.

This makes default compilation ~30% faster,
[clang, make -j8, old]: real    1m04.355s
[clang, make -j8, new]: real    0m41.740s

For new case, test needs to built separately, which takes,
[clang, make -j8 test]: real    0m24.293s

The point is tests are not required always and by every one.

Signed-off-by: Ferruh Yigit <ferruh.yigit@intel.com>