crypto/scheduler: add failover scheduling mode
This patch adds the failover scheduling mode to scheduler crypto
PMD. In this mode the 1st and 2nd slaves are used as primary and
secondary slaves respectively. The scheduler will try to enqueue
the incoming crypto operations to the primary slave first, and
will only enqueue to the secondary slave upon the firt enqueue
is failed.

Signed-off-by: Fan Zhang <roy.fan.zhang@intel.com>