examples/performance-thread: remove reference to execinfo.h
There is no function to refer any part of execinfo.h, so remove the
reference to it. And there is no this file in musl. So need to
remove it to support musl.

The musl is an alternative LIBC to GLIBC and provides the
standard C/POSIX library and extensions.
The musl can be got from http://www.musl-libc.org

Signed-off-by: Wei Dai <wei.dai@intel.com>